FaceMesh3D/requirements.txt
# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Deep learning
torch>=2.2.0
torchvision>=0.17.0

# Face detection (used by DECA internally)
face-alignment>=1.3.5

# Image processing
Pillow>=10.2.0
numpy>=1.26.0
scikit-image>=0.22.0
opencv-python-headless>=4.9.0

# 3D / mesh utilities
trimesh>=4.3.0
PyOpenGL>=3.1.7

# DECA / FLAME dependencies
chumpy>=0.70
scipy>=1.12.0
